lint: address detekt issues
