+++ .idea/gradle.xml
@@ -17,6 +17,9 @@
             <option value="$PROJECT_DIR$/database/models" />
             <option value="$PROJECT_DIR$/features" />
             <option value="$PROJECT_DIR$/features/albums" />
+            <option value="$PROJECT_DIR$/features/albums/create" />
+            <option value="$PROJECT_DIR$/features/albums/home" />
+            <option value="$PROJECT_DIR$/features/albums/shared" />
             <option value="$PROJECT_DIR$/platform" />
             <option value="$PROJECT_DIR$/platform/error" />
             <option value="$PROJECT_DIR$/platform/injection" />
