+++ features/albums/create/src/main/kotlin/features/albums/create/di/CreateAlbumStepDi.kt
@@ -1,0 +1,12 @@
+package features.albums.create.di
+
+import cafe.adriel.voyager.core.registry.screenModule
+import features.albums.create.presentation.NameScreen
+import platform.navigation.NavigationProvider
+
+val createAlbumStepDi = screenModule {
+    register<NavigationProvider.Albums.Create> {
+        NameScreen
+    }
+
+}
