+++ features/albums/create/src/main/kotlin/features/albums/create/presentation/CreateAlbumState.kt
@@ -1,0 +1,8 @@
+package features.albums.create.presentation
+
+internal data class CreateAlbumState(
+    val name: String = "",
+    val description: String = "",
+) {
+    val isButtonEnabled = name.isNotEmpty() && description.isNotEmpty()
+}
