+++ features/albums/home/src/main/kotlin/features/albums/home/di/albumHomeStepDi.kt
@@ -1,0 +1,11 @@
+package features.albums.home.di
+
+import cafe.adriel.voyager.core.registry.screenModule
+import features.albums.home.presentation.HomeScreen
+import platform.navigation.NavigationProvider
+
+val albumHomeStepDi = screenModule {
+    register<NavigationProvider.Albums.Home> {
+        HomeScreen
+    }
+}
