+++ features/albums/home/src/main/kotlin/features/albums/home/presentation/HomeState.kt
@@ -1,0 +1,8 @@
+package features.albums.home.presentation
+
+import database.models.models.AlbumModel
+
+internal data class HomeState(
+    val isInitialized: Boolean = false,
+    val albums: List<AlbumModel> = listOf(),
+)
