+++ features/albums/shared/src/main/kotlin/features/albums/shared/domain/model/CreateAlbumDTO.kt
@@ -1,0 +1,6 @@
+package features.albums.shared.domain.model
+
+data class CreateAlbumDTO(
+    val name: String,
+    val description: String,
+)
