--- features/albums/src/main/kotlin/br/com/alaksion/features/albums/domain/model/CreateAlbumDTO.kt
@@ -1,6 +1,0 @@
-package br.com.alaksion.features.albums.domain.model
-
-internal data class CreateAlbumDTO(
-    val name: String,
-    val description: String,
-)
