--- features/albums/src/main/kotlin/br/com/alaksion/features/albums/presentation/create/CreateAlbumState.kt
@@ -1,8 +1,0 @@
-package br.com.alaksion.features.albums.presentation.create
-
-internal data class CreateAlbumState(
-    val name: String = "",
-    val description: String = "",
-) {
-    val isButtonEnabled = name.isNotEmpty() && description.isNotEmpty()
-}
