--- features/albums/src/main/kotlin/br/com/alaksion/features/albums/presentation/create/CreateViewModel.kt
@@ -1,73 +1,0 @@
-package br.com.alaksion.features.albums.presentation.create
-
-import br.com.alaksion.features.albums.domain.model.CreateAlbumDTO
-import br.com.alaksion.features.albums.domain.repository.AlbumRepository
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.CoroutineDispatcher
-import platform.injection.IODispatcher
-import platform.uistate.uievent.UiEvent
-import platform.uistate.uievent.UiEventHandler
-import platform.uistate.uievent.UiEventHandlerImpl
-import platform.uistate.uistate.UiStateViewModel
-import java.util.UUID
-import javax.inject.Inject
-
-internal sealed class CreateAlbumIntent {
-    data class UpdateName(val value: String) : CreateAlbumIntent()
-    data class UpdateDescription(val value: String) : CreateAlbumIntent()
-    object CreateAlbum : CreateAlbumIntent()
-}
-
-internal sealed class CreateAlbumEvents(val result: AlbumResult) : UiEvent {
-    class Result(
-        result: AlbumResult,
-        override val eventId: UUID
-    ) : CreateAlbumEvents(result)
-}
-
-@HiltViewModel
-internal class CreateViewModel @Inject constructor(
-    @IODispatcher dispatcher: CoroutineDispatcher,
-    private val repository: AlbumRepository
-) : UiStateViewModel<CreateAlbumState>(CreateAlbumState(), dispatcher),
-    UiEventHandler<CreateAlbumEvents> by UiEventHandlerImpl() {
-
-    fun handleIntent(intent: CreateAlbumIntent) {
-        when (intent) {
-            is CreateAlbumIntent.UpdateName -> updateName(intent.value)
-            is CreateAlbumIntent.UpdateDescription -> updateDescription(intent.value)
-            CreateAlbumIntent.CreateAlbum -> createAlbum()
-        }
-    }
-
-    private fun updateName(value: String) {
-        setState(showLoading = false) { currentState ->
-            currentState.copy(name = value)
-        }
-    }
-
-    private fun updateDescription(value: String) {
-        setState(showLoading = false) { currentState ->
-            currentState.copy(description = value)
-        }
-    }
-
-    private fun createAlbum() {
-        runSuspendCatching {
-            val result = kotlin.runCatching {
-                repository.createAlbum(
-                    data = CreateAlbumDTO(
-                        name = stateData.name,
-                        description = stateData.description
-                    )
-                )
-            }.fold(
-                onSuccess = { CreateAlbumEvents.Result(AlbumResult.Success, UUID.randomUUID()) },
-                onFailure = { CreateAlbumEvents.Result(AlbumResult.Error, UUID.randomUUID()) },
-            )
-
-            enqueueEvent(result)
-        }
-    }
-
-}
