--- features/albums/src/main/kotlin/br/com/alaksion/features/albums/presentation/home/HomeState.kt
@@ -1,8 +1,0 @@
-package br.com.alaksion.features.albums.presentation.home
-
-import database.models.models.AlbumModel
-
-internal data class HomeState(
-    val isInitialized: Boolean = false,
-    val albums: List<AlbumModel> = listOf(),
-)
