+++ includes.gradle.kts
@@ -5,4 +5,6 @@ include(":platform:uicomponents")
 include(":platform:navigation")
 include(":platform:injection")
 include(":platform:error")
-include(":features:albums")
+include(":features:albums:home")
+include(":features:albums:create")
+include(":features:albums:shared")
